fix: use wagmi core methods
